requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
bokeh==3.6.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
colorcet==3.1.0
colorlog==6.9.0
contourpy==1.3.2
cucim-cu12==25.4.0
cuda-bindings==12.9.0
cuda-python==12.9.0
cudf-cu12==25.4.0
cugraph-cu12==25.4.1
cuml-cu12==25.4.0
cuproj-cu12==25.4.0
cupy-cuda12x==13.4.1
cuspatial-cu12==25.4.0
cuvs-cu12==25.4.0
cuxfilter-cu12==25.4.0
dask==2025.2.0
dask-cuda==25.4.0
dask-cudf-cu12==25.4.0
datashader==0.18.1
defusedxml==0.7.1
distributed==2025.2.0
distributed-ucxx-cu12==0.43.0
fastjsonschema==2.21.1
fastrlock==0.8.3
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
geopandas==1.0.1
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
holoviews==1.20.2
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
isoduration==20.11.0
Jinja2==3.1.6
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
lazy_loader==0.4
libcudf-cu12==25.4.0
libcugraph-cu12==25.4.1
libcuml-cu12==25.4.0
libcuspatial-cu12==25.4.0
libcuvs-cu12==25.4.0
libkvikio-cu12==25.4.0
libraft-cu12==25.4.0
librmm-cu12==25.4.0
libucx-cu12==1.18.1
libucxx-cu12==0.43.0
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
msgpack==1.1.0
multidict==6.4.3
multipledispatch==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numba==0.60.0
numba-cuda==0.4.0
numpy==2.0.2
nvidia-cublas-cu12==12.9.0.13
nvidia-cufft-cu12==11.4.0.6
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.4.40
nvidia-cusparse-cu12==12.5.9.5
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.5
nvidia-nvcomp-cu12==4.2.0.11
nvidia-nvjitlink-cu12==12.9.41
nvtx==0.2.11
nx-cugraph-cu12==25.4.0
optuna==4.3.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
panel==1.6.3
param==2.2.0
partd==1.4.2
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.21.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0rc2
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyct==0.5.0
Pygments==2.19.1
pylibcudf-cu12==25.4.0
pylibcugraph-cu12==25.4.1
pylibraft-cu12==25.4.0
pynvjitlink-cu12==0.6.0
pynvml==12.0.0
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.4.0
raft-dask-cu12==25.4.0
rapids-dask-dependency==25.4.0
rapids-logger==0.1.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rmm-cu12==25.4.0
rpds-py==0.24.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
shapely==2.0.7
simpervisor==1.0.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
tblib==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treelite==4.4.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
ucx-py-cu12==0.43.0
ucxx-cu12==0.43.0
uri-template==1.3.0
urllib3==2.4.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xarray==2025.4.0
xyzservices==2025.4.0
yarl==1.20.0
zict==3.0.0
